Added Health System and Attacking
When the player attacks, it now checks to see if it collided with an enemy. If so, it applies damage to them. The Player Damage stuff is inside 'HatScript' script file in the Player Scripts folder if you want to look at the code. Also changed the EnemyBase GameObject in ArtTestZone scene to the EnemyBase Prefab object since it wasn't associated with the prefab.