Added Enemy States and Perception
Enemy now can move between patrol nodes / transforms and when it reaches a node, it switches states to Idle. Idle is pretty much just a timer right now that waits for a certain amount of seconds and then changes the target object to the next patrol node and switches back to the Move state.

The enemy also has an EnemyPerception script that checks if the player is nearby and in sight. Currently doesn't do anything right now.