AI Improvements
AI now goes back to non aggro if the target object is out of sight for a certain amount of time. Also, enemy now does it's entire attack instead of just stopping in the middle of its attack when the player moves out of reach. The sprite also flips based on which direction it is moving in as well.