Enemy AI Template
Template for the enemy AI is fully functional. Enemy moves between move points, and when it spots the player, it moves towards the player. If the player is out of sight from the enemy, then the enemy will go into its idle state and after a certain amount of time the player has been out of sight, the enemy will go back to it's non aggro behavior