GorbAI But Better
Added movement stats to the gorb, and also calls to variables to use the gorb stats inside the esm so its easier to change which stats your using. I tried to make this easier but it's not worth the time right now. Also added the fall state. The enemy falls correctly now, but the animator needs fixed because I hooked it up wrong.