Rough Draft for Enemy Knockback
Enemy knockback is implemented, though the switching of substates is a little weird, so that needs fixed. But the knockback works. Also, if you want to change the values for how much knockback should effect the enemy, you can change the values in the Enemy Movement Stats or Gorb Movement stats (there are different stats for each enemy)