Volley Set-Up When Sliding
When the player slides into an enemy, the enemy now moves up and slightly behind the player as we talked about. There are a few tiny bugs where the enemy sometimes doesn't have the right velocity when they are off the ground, but hopefully that will be fixed when volley is in. Also, I'm using the names of objects to decide the velocity of knockback and whether the enemy was slided into, so it will probably be needed to change.